Time: 3 ms (88.76%) | Memory: 9 MB (8.30%) - LeetSync

+++ 201-bitwise-and-of-numbers-range/bitwise-and-of-numbers-range.cpp
@@ -4,7 +4,7 @@ public:
         if(left==right) return left;
         if(left==0) return 0;
         if(__bit_width(left) != __bit_width(right)) return 0;
-        int_fast8_t count=0;
+        short count=0;
         while(left != right){
             count++;
             left>>=1;
